% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{$.qualtrics}
\alias{$.qualtrics}
\title{Subset a \code{qualtrics} data frame}
\usage{
\method{$}{qualtrics}(x, name)
}
\arguments{
\item{x}{a data frame of class \code{qualtrics}}

\item{name}{a character string used for extraction. This is matched to the names of the object to subset the desired columns}
}
\description{
Given an object of class \code{qualtrics}, this \code{\link[base:Extract]{base::Extract}} subsetting method allows the user to subset a data frame while preserving its labels.
}
