% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Results from a sample survey downloaded from Qualtrics}
\format{
A data frame with 102 rows and 42 columns
\describe{
\item{start_date}{The date the participant started the survey}
\item{end_date}{The date the participant finished the survey}
\item{status}{???}
\item{ip_address}{The IP address of a participant's device}
\item{progress}{How far the participant got into the survey}
\item{duration_in_seconds}{How many seconds it took for the participant to complete the survey?}
\item{finished}{A logical response indicating if the participant completed the survey or not}
\item{recorded_date}{The date the participant's response was recorded}
\item{resonse_id}{The unique ID assigned to each participant's response}
\item{recipient_last_name}{The last name provided by the participant}
\item{recipient_first_name}{The first name provided by the participant}
\item{recipient_email}{The email provided by the participant}
\item{external_reference}{?????}
\item{location_latitude}{The latidude where the survey response was taken}
\item{location_longitude}{The longitude where the survey response was taken}
\item{distribution_channel}{The method through which the survey was administered ?????}
\item{user_language}{The language the participant took the survey in}
\item{elig_1}{Eligibiltiy: meal delivery services used within the last 3 months}
\item{csat}{Satisfaction with product or service}
\item{churn}{Likelihood to switch to a new brand}
\item{drivers_1}{Statement 1: Satisfaction with time required to cook}
\item{drivers_2}{Statement 2: Satisfaction with ease of cooking instructions}
\item{drivers_3}{Statement 3: Satisfaction with quality of ingredients}
\item{drivers_4}{Statement 4: Satisfaction with value for the cost}
\item{drivers_5}{Statement 5: Satisfaction with website experience}
\item{drivers_6}{Statement 6: Satisfaction with taste/flavor of meals}
\item{drivers_7}{Statement 7: Satisfaction with variety over time}
\item{drivers_8}{Statement 8: Satisfaction with portion size}
\item{drivers_9}{Statement 9: Satisfaction with availability of offerings that meet participant's nutritional needs}
\item{drivers_10}{Statement 10: Satisfaction with shipping & packaging}
\item{age}{Age of the participant}
\item{gender}{Participant's gender identity}
\item{gender_4_text}{Participant's gender identity self description}
\item{education}{Highest level of education participant completed}
\item{employment}{Participant's employment status}
\item{income}{Participant's total household income before taxes in the past year}
\item{marital}{Participant's marital status}
\item{spanish_hispanic_latino}{Participant of Spanish, Hispanic, or Latino orgin}
\item{race}{Race of participant}
\item{household}{Number of people living/staying in participant's household at least half of the time}
\item{zip}{Participant's zip code}
\item{q_state_zip}{Abbreviation of state associated with participant's zip code}
}
}
\usage{
example_data
}
\description{
Results from a sample survey downloaded from Qualtrics
}
\keyword{datasets}
