% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{get_questions}
\alias{get_questions}
\title{Survey questions}
\usage{
get_questions(data)
}
\arguments{
\item{data}{A data frame to be tidied}
}
\value{
A list of all the survey's questions as character vectors
}
\description{
This function takes the first row of an unedited Qualtrics dataframe creates a new object of class
\code{list} that contains every question in the survey in the form of character vectors.
}
