% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\docType{data}
\name{heuristics}
\alias{heuristics}
\title{Qualtrics Likert-scale presets (re-class heuristics)}
\format{
An object of class \code{list} of length 25.
}
\usage{
heuristics
}
\description{
This list contains the Qualtrics Likert scale preset values. When the
\code{\link[=make_likert]{make_likert()}} function is called, this is the list of possible Likert scales
it references when attempting to re-class the data frame.
}
\keyword{datasets}
