% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{is_qualtrics}
\alias{is_qualtrics}
\title{Check if data frame is in unedited Qualtrics format}
\usage{
is_qualtrics(data)
}
\arguments{
\item{data}{Data attempting to be tidied}
}
\description{
This function checks if the data to be tidied is in the format of unedited Qualtrics
data. If the data does not appear to be be from Qualtrics, it will signal a condition.
}
