% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{make_likert}
\alias{make_likert}
\title{Re-class columns with Qualtrics likert-scale presets}
\usage{
make_likert(data)
}
\arguments{
\item{data}{Data frame to iterate over in search of Likert scale responses}
}
\value{
Returns the same data frame updated with re-classed Likert scale columns
Will display a message listing which (if any) columns were re-classed
}
\description{
This function iterates over each column of class character in a data frame and
compares its unique responses to Qualtrics preset Likert scale labels, called
\code{\link{heuristics}}. If the responses correspond with a given heuristic, the function
will re-lass the column as a factor with the same levels as the heuristic.
}
