% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{new_qualtrics}
\alias{new_qualtrics}
\title{Create an object of class \code{qualtrics}}
\usage{
new_qualtrics(data)
}
\arguments{
\item{data}{A data set exported from Qualtrics}
}
\value{
An object of class \code{qualtrics} and \code{data.frame}
}
\description{
Given a data frame object, this function adds the class \code{qualtrics} to the object
}
