% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{Check if data is in Qualtrics format}
\usage{
read_qualtrics(path, heuristics = TRUE)
}
\arguments{
\item{path}{File path to csv being imported}

\item{heuristics}{Logical to determine if function should try identify and class likert questions}
}
\description{
This function reads a csv file according to the standard Qualtric format
}
