% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{[[.qualtrics}
\alias{[[.qualtrics}
\title{Subset a \code{qualtrics} data frame}
\usage{
\method{[[}{qualtrics}(x, ..., exact = TRUE)
}
\arguments{
\item{x}{a data frame of class \code{qualtrics}}

\item{...}{additional arguments used by \code{[[}}

\item{exact}{set to default to \code{TRUE}, controls possible partial matching when extracting by a character vector}
}
\description{
Given an object of class \code{qualtrics}, this \code{\link[base:Extract]{base::Extract}} subsetting method allows the user to subset a data frame while preserving its labels.
}
