% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy Qualtrics data frame}
\usage{
tidy_data(data, force = FALSE)
}
\arguments{
\item{data}{A data frame to be tidied}

\item{force}{Logical to determine if warnings can be overruled}
}
\value{
The tidied data frame and a list of all the questions as character vectors
}
\description{
This function puts data imported from Qualtrics into "tidy" format by removing
the first two rows, as they do not contain actual observations. In addition,
it takes the first of these unnecessary rows and creates a new object of class
\code{list} that contains every question in the survey in the form of character vectors.
}
