% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{validate_qualtrics}
\alias{validate_qualtrics}
\title{Validate an object of class \code{qualtrics}}
\usage{
validate_qualtrics(data)
}
\value{
A data frame of class \code{qualtrics}
}
\description{
Given a object of class \code{qualtrics}, this function will check that
the object satisfies all of the conditions necessary to be an object of class \code{qualtrics}.
It checks:
\itemize{
\item if the object is a data frame
\item if the data frame has labels
\item if the data frame's first two columns are named "start_date" and "end_date"
}
}
