% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Standardize column names}
\usage{
clean_names(data)
}
\arguments{
\item{data}{The data frame containing the column names to be cleaned.}
}
\value{
a data frame with column names in snake case.
}
\description{
This function takes column names and tidies them, returning the names in snake case.
}
