% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{Imports and clean Qualtrics csv data}
\usage{
read_qualtrics(path, heuristics = TRUE)
}
\arguments{
\item{path}{File path to csv being imported}

\item{heuristics}{Logical to determine if function should try identify and class likert questions}
}
\description{
This function reads a csv file according to the standard Qualtric format.
}
\note{
During importation, the question text is saved as a label for each column.
To access this label use \code{\link[labelVector:get_label]{labelVector::get_label()}}.
}
\examples{
file_path <- fs::path_package("extdata", "qualtrics_ex_data.csv", package = "EasyQualtrics")
mydata <- read_qualtrics(file_path)


}
