% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{[.qualtrics}
\alias{[.qualtrics}
\title{Subset a \code{qualtrics} data frame}
\usage{
\method{[}{qualtrics}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{a data frame of class \code{qualtrics}}

\item{i}{typically used for row indexing}

\item{j}{typically use for column indexing}

\item{...}{additional arguments used by \code{[}}

\item{drop}{set to default to \code{FALSE}, preserves the data frame's dimensions}
}
\description{
Given an object of class \code{qualtrics}, this \code{\link[base:Extract]{base::Extract}} subsetting method allows the user to subset a data frame while preserving its labels.
}
\note{
This subsetting method does not allow for dimensions to be dropped. Thus, to
subset a single column of a data frame use either \code{$} or \code{[[}.
}
