% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{[[.qualtrics}
\alias{[[.qualtrics}
\title{Subset a \code{qualtrics} data frame}
\usage{
\method{[[}{qualtrics}(data, ...)
}
\arguments{
\item{data}{a data frame of class \code{qualtrics}}

\item{...}{additional arguments used by \code{[[}}
}
\description{
Given an object of class \code{qualtrics}, this subsetting method allows the user to subset a data frame while preserving its labels.
}
