% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{[[.qualtrics}
\alias{[[.qualtrics}
\title{Subset a \code{qualtrics} data frame}
\usage{
\method{[[}{qualtrics}(x, i, j, ...)
}
\arguments{
\item{...}{additional arguments for \code{[[}}

\item{data}{a data frame of class \code{qualtrics}}
}
\description{
Given an object of class \code{qualtrics}, this \code{[[} subsetting method allows the user to subset a data frame while preserving its labels.}
